Refactor Gemini test layout
